refactor(mpp): extract shape-generic exec_mpp_worker into mpp::exec_worker
The natural-shape worker dispatcher (deserialize logical plan, build
distributed physical plan, find this proc's fragments, run them via
join_all) is identical for every customscan that opts into MPP. Only
the seed SessionContext (aggregate vs join profile) and where the
inputs come from in per-scan state are customscan-specific.

Pull the dispatcher out into mpp::exec_worker:
- MppWorkerInputs bundles the per-call inputs.
- build_mpp_session_context(seed, mesh) layers the distributed
  planner knobs over an arbitrary seed context.
- run_mpp_worker(inputs, seed, runtime) is the shape-agnostic
  dispatch loop.

AggregateScan::exec_mpp_worker becomes a thin wrapper: extracts
inputs from its typed state, calls create_aggregate_session_context
for the seed, hands off to run_mpp_worker. The leader-side
AggregateScan::build_mpp_session_context is now a one-liner over the
shared helper.

No behavior change. Sets up wiring JoinScan through the same dispatcher.
